requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
blas=1.1=openblas
bzip2=1.0.6=h14c3975_1002
ca-certificates=2018.11.29=ha4d7672_0
cairo=1.14.12=h8948797_3
certifi=2018.11.29=py36_1000
cffi=1.11.5=py36he75722e_1
cudatoolkit=9.0=h13b8566_0
cudnn=7.3.1=cuda9.0_0
cycler=0.10.0=py36_0
dbus=1.13.6=h746ee38_0
expat=2.2.6=he6710b0_0
ffmpeg=4.1=h6dce934_1001
fontconfig=2.13.0=h9420a91_0
freetype=2.9.1=h8a8886c_1
giflib=5.1.4=h14c3975_1001
glib=2.56.2=hd408876_0
gmp=6.1.2=hf484d3e_1000
gnutls=3.6.5=hd3a4fd2_1001
graphite2=1.3.13=hf484d3e_1000
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb453b48_1
harfbuzz=1.9.0=he243708_1001
hdf5=1.10.4=nompi_h11e915b_1105
icu=58.2=h9c2bf20_1
intel-openmp=2019.1=144
jasper=1.900.1=h07fcdf6_1006
jpeg=9c=h14c3975_1001
kiwisolver=1.0.1=py36hf484d3e_0
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc=7.2.0=h69d50b8_2
libgcc-ng=8.2.0=hdf63c60_1
libgfortran-ng=7.3.0=hdf63c60_0
libiconv=1.15=h14c3975_1004
libopenblas=0.3.3=h5a2b251_3
libpng=1.6.36=hbc83047_0
libstdcxx-ng=8.2.0=hdf63c60_1
libtiff=4.0.10=h2733197_2
libuuid=1.0.3=h1bed415_2
libwebp=1.0.2=h576950b_1
libxcb=1.13=h1bed415_1
libxml2=2.9.9=he19cac6_0
matplotlib=3.0.2=py36h8a2030e_1001
matplotlib-base=3.0.2=py36h167e16e_1001
mkl=2018.0.3=1
mkl_fft=1.0.10=py36_0
mkl_random=1.0.1=py36h629b387_0
nccl=1.3.5=cuda9.0_0
ncurses=6.1=he6710b0_1
nettle=3.4.1=h14c3975_1002
ninja=1.8.2=py36h6bb024c_1
numpy=1.16.1=py36_blas_openblash1522bff_0
numpy-base=1.15.4=py36h2f8d375_0
olefile=0.46=py36_0
openblas=0.3.3=h9ac9557_1001
opencv=3.4.4=py36_blas_openblash85ad109_1203
openh264=1.8.0=hdbcaa40_1000
openssl=1.1.1a=h14c3975_1000
pcre=8.42=h439df22_0
pillow=5.4.1=py36h34e0f95_0
pip=19.0.1=py36_0
pixman=0.34.0=h14c3975_1003
pycparser=2.19=py36_0
pyparsing=2.3.1=py36_0
pyqt=5.6.0=py36_2
python=3.6.8=h0371630_0
python-dateutil=2.7.5=py36_0
pytorch=0.4.0=py36hdf912b8_0
pytz=2018.9=py36_0
qt=5.6.3=h8bf5577_3
readline=7.0=h7b6447c_5
scipy=1.2.1=py36_blas_openblash1522bff_0
setuptools=40.8.0=py36_0
sip=4.19.8=py36hf484d3e_0
six=1.12.0=py36_0
sqlite=3.26.0=h7b6447c_0
tk=8.6.8=hbc83047_0
torchvision=0.2.1=py36_0
tornado=5.1.1=py36h7b6447c_0
wheel=0.32.3=py36_0
x264=1!152.20180717=h14c3975_1001
xz=5.2.4=h14c3975_4
zlib=1.2.11=h7b6447c_3
zstd=1.3.7=h0b5b093_0
